requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
httpx==0.28.1
